requirements.txt
attrs==23.2.0
cairocffi==1.7.1
CairoSVG==2.7.1
cffi==1.17.1
cssselect2==0.7.0
defusedxml==0.7.1
deprecation==2.1.0
lottie==0.7.1
mysql-connector-python==9.1.0
numpy==1.26.4
packaging==24.2
pillow==10.4.0
pycparser==2.22
pyrlottie==2024.0.1
pyserial==3.5
python-dotenv==1.0.1
rlottie-python==1.3.6
RPi.GPIO==0.7.1
spidev==3.6
tinycss2==1.4.0
webencodings==0.5.1
